+++ unkey/routes.py
@@ -80,7 +80,7 @@ class Route:
 
 
 # Keys
-CREATE_KEY: t.Final[Route] = Route(c.POST, "/keys")
+CREATE_KEY: t.Final[Route] = Route(c.POST, "/keys.createKey")
 VERIFY_KEY: t.Final[Route] = Route(c.POST, "/keys/verify")
 REVOKE_KEY: t.Final[Route] = Route(c.DELETE, "/keys/{}")
 UPDATE_KEY: t.Final[Route] = Route(c.PUT, "/keys/{}")